+++ ddtrace/llmobs/_llmobs.py
@@ -1,4 +1,3 @@
-import inspect
 import json
 import os
 import sys
@@ -46,6 +45,13 @@ from ddtrace.llmobs._constants import SPAN_KIND
 from ddtrace.llmobs._constants import SPAN_START_WHILE_DISABLED_WARNING
 from ddtrace.llmobs._constants import TAGS
 from ddtrace.llmobs._evaluators.runner import EvaluatorRunner
+from ddtrace.llmobs._links import TrackedDict
+from ddtrace.llmobs._links import TrackedList
+from ddtrace.llmobs._links import TrackedStr
+from ddtrace.llmobs._links import add_span_links
+from ddtrace.llmobs._links import get_obj_id
+from ddtrace.llmobs._links import search_links_from_relationships
+from ddtrace.llmobs._links import track_object_interactions
 from ddtrace.llmobs._trace_processor import LLMObsTraceProcessor
 from ddtrace.llmobs._utils import AnnotationContext
 from ddtrace.llmobs._utils import _get_llmobs_parent_id
@@ -72,147 +78,6 @@ SUPPORTED_LLMOBS_INTEGRATIONS = {
     "langchain": "langchain",
     "google_generativeai": "google_generativeai",
 }
-
-
-def track_object_interactions(frame, event, arg):
-    """Track method calls and object interactions."""
-    if event != "call":
-        return track_object_interactions
-
-    # Get function details
-    source_obj = frame.f_locals.get("self")
-    method_name = frame.f_code.co_name
-
-    # Track object creation or modification
-    try:
-        # Capture arguments that might involve object relationships
-        args = inspect.getargvalues(frame)
-
-        # Look for potential object interactions
-        for arg_name in args.args:
-            if arg_name == "self":
-                continue
-            incoming_obj = frame.f_locals.get(arg_name)
-
-            # Check for object creation or modification
-            if method_name in [
-                "__init__",
-                "__new__",
-                "__add__",
-                "append",
-                "extend",
-                "update",
-            ]:
-                LLMObs._instance._record_relationship(source_obj, incoming_obj, f"{method_name} operation")
-    except Exception as e:
-        print("Error capturing object interactions ", e)
-
-    return track_object_interactions
-
-
-class TrackedStr(str):
-    def __add__(self, other):
-        result = super().__add__(other)
-        result = TrackedStr(result)
-        LLMObs._instance._record_relationship(result, other, "__add__")
-        LLMObs._instance._record_relationship(result, self, "__add__")
-        return result
-
-    def __radd__(self, other):
-        result = super().__radd__(other)
-        result = TrackedStr(result)
-        LLMObs._instance._record_relationship(result, other, "__radd__")
-        LLMObs._instance._record_relationship(result, self, "__radd__")
-        return result
-
-    def format(self, *args, **kwargs):
-        result = super().format(*args, **kwargs)
-        result = TrackedStr(result)
-        for arg in args:
-            LLMObs._instance._record_relationship(result, arg, "format")
-        for key, value in kwargs.items():
-            LLMObs._instance._record_relationship(result, value, "format")
-        LLMObs._instance._record_relationship(result, self, "format")
-        return result
-
-    def split(self, *args, **kwargs):
-        return super().split(*args, **kwargs)
-
-    def join(self, *args, **kwargs):
-        return super().join(*args, **kwargs)
-
-
-class TrackedList(list):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        for item in self:
-            LLMObs._instance._record_relationship(self, item, "__init__")
-
-    def append(self, item):
-        result = super().append(item)
-        LLMObs._instance._record_relationship(self, item, "append")
-        return result
-
-    def extend(self, iterable):
-        result = super().extend(iterable)
-        LLMObs._instance._record_relationship(self, iterable, "extend")
-        return result
-
-    def __delitem__(self, key) -> None:
-        if key < len(self):
-            LLMObs._instance._remove_relationship(self, self[key], "__delitem__")
-        super().__delitem__(key)
-
-    def __setitem__(self, key, value):
-        super().__setitem__(key, value)
-        LLMObs._instance._record_relationship(self, value, "__setitem__")
-
-    def __add__(self, other):
-        result = super().__add__(other)
-        LLMObs._instance._record_relationship(self, other, "__add__")
-        return result
-
-
-class TrackedDict(dict):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        for key, value in self.items():
-            LLMObs._instance._record_relationship(self, key, "__init__")
-            LLMObs._instance._record_relationship(self, value, "__init__")
-
-    def fromkeys(self, *args, **kwargs):
-        ret = super().fromkeys(*args, **kwargs)
-        for key, value in ret.items():
-            LLMObs._instance._record_relationship(ret, key, "fromkeys")
-            LLMObs._instance._record_relationship(ret, value, "fromkeys")
-        return ret
-
-    def update(self, *args, **kwargs):
-        result = super().update(*args, **kwargs)
-        for arg in args:
-            LLMObs._instance._record_relationship(self, arg, "update")
-        return result
-
-    def pop(self, key, *args, **kwargs):
-        val = super().pop(key, *args, **kwargs)
-        LLMObs._instance._remove_relationship(self, key, "pop")
-        if val is not None:
-            LLMObs._instance._remove_relationship(self, val, "pop")
-        return val
-
-    def __delitem__(self, key) -> None:
-        val = self.get(key)
-        super().__delitem__(key)
-        LLMObs._instance._remove_relationship(self, key, "__delitem__")
-        if val is not None:
-            LLMObs._instance._remove_relationship(self, val, "__delitem__")
-
-    def __setitem__(self, key, value):
-        if key in self:
-            LLMObs._instance._remove_relationship(self, self[key], "__setitem__")
-        super().__setitem__(key, value)
-        LLMObs._instance._record_relationship(self, value, "__setitem__")
-        LLMObs._instance._record_relationship(self, key, "__setitem__")
 
 
 class LLMObs(Service):
@@ -249,8 +114,6 @@ class LLMObs(Service):
         self._annotation_context_lock = forksafe.RLock()
         self.tracer.on_start_span(self._do_annotations)
 
-        self._object_span_links = {}
-        self._object_relationships = {}
         sys.settrace(track_object_interactions)
 
     def _do_annotations(self, span):
@@ -416,35 +279,6 @@ class LLMObs(Service):
 
         log.debug("%s disabled", cls.__name__)
 
-    def _remove_relationship(self, source_obj, incoming_obj, operation):
-        """Record a relationship between objects."""
-        if source_obj is None or incoming_obj is None:
-            return
-        if source_obj not in self._object_relationships:
-            self._object_relationships[self.get_obj_id(source_obj)] = set()
-        self._object_relationships[self.get_obj_id(source_obj)].remove(self.get_obj_id(incoming_obj))
-
-    def _record_relationship(self, source_obj, incoming_obj, operation):
-        """Record a relationship between objects."""
-        if source_obj is None or incoming_obj is None:
-            return
-        if self.get_obj_id(source_obj) not in self._object_relationships:
-            self._object_relationships[self.get_obj_id(source_obj)] = set()
-        self._object_relationships[self.get_obj_id(source_obj)].add(self.get_obj_id(incoming_obj))
-
-    def get_all_links(self, obj_id, visited: Optional[set] = None, links: Optional[list] = None):
-        if visited is None:
-            visited = set()
-        visited.add(obj_id)
-        if links is None:
-            links = []
-        if obj_id not in self._object_relationships:
-            return self._object_span_links.get(obj_id, [])
-        for obj in self._object_relationships[obj_id]:
-            if obj not in visited:
-                links += self.get_all_links(obj, visited, links)
-        return self._object_span_links.get(obj_id, [])
-
     def record_object(self, span, obj, to):
         if isinstance(obj, str) and not isinstance(obj, TrackedStr):
             obj = TrackedStr(obj)
@@ -454,14 +288,14 @@ class LLMObs(Service):
             obj = TrackedDict(obj)
 
         current_span_links = []
-        for span_link in self.get_all_links(self.get_obj_id(obj)):
+        for span_link in search_links_from_relationships(get_obj_id(obj)):
             if span_link["attributes"]["from"] == "input" and to == "output":
                 continue
             span_link["attributes"]["to"] = to
             current_span_links.append(span_link)
 
         self._tag_span_links(span, current_span_links)
-        self.add_span_links(
+        add_span_links(
             obj,
             [
                 {
@@ -484,18 +318,6 @@ class LLMObs(Service):
             current_span_links = json.loads(current_span_links)
             span_links = current_span_links + span_links
         span.set_tag_str("_ml_obs.span_links", safe_json(span_links))
-
-    def get_obj_id(self, obj):
-        return f"{type(obj).__name__}_{id(obj)}"
-
-    def add_span_links(self, obj, span_links):
-        obj_id = self.get_obj_id(obj)
-        if obj_id not in self._object_span_links:
-            self._object_span_links[obj_id] = []
-        self._object_span_links[obj_id] += span_links
-
-    def get_span_links(self, obj):
-        return self._object_span_links.get(self.get_obj_id(obj), [])
 
     @classmethod
     def annotation_context(
